requirements.txt
python-binance==1.0.19
typer==0.12.5
rich==13.9.2
httpx==0.27.2
python-dotenv==1.0.1
fastapi==0.115.5
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
Jinja2==3.1.4
pytest==8.3.4